refactor: extract common SSE logic (#37)
# Conflicts:
#	src/main/java/com/kt/mindLog/service/session/SessionStreamService.java
