+++ src/main/java/com/kt/mindLog/service/session/SessionStreamService.java
@@ -26,7 +26,7 @@ import com.kt.mindLog.global.common.support.Preconditions;
 import com.kt.mindLog.global.property.SessionProperties;
 import com.kt.mindLog.repository.UserRepository;
 import com.kt.mindLog.repository.session.SessionRepository;
-import com.kt.mindLog.service.summary.SummaryService;
+import com.kt.mindLog.service.sse.SSEService;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -42,40 +42,17 @@ public class SessionStreamService {
 	private final SessionRepository sessionRepository;
 
 	private final ObjectMapper objectMapper;
-	private final WebClient webClient;
 	private final SessionProperties sessionProperties;
 
+	private final SSEService sseService;
 	private final SessionMessageService messageService;
-
-	//SSE (공통)
-	private Flux<ServerSentEvent<String>> streamSse(final String uri, final UUID sessionId, final Object body) {
-		var request = webClient.post()
-			.uri(uri, sessionId)
-			.contentType(MediaType.APPLICATION_JSON)
-			.accept(MediaType.TEXT_EVENT_STREAM);
-
-		if (body != null) {
-			request = (WebClient.RequestBodySpec)request.bodyValue(body);
-		}
-
-		return request
-			.retrieve()
-			.onStatus(HttpStatusCode::isError, response ->
-				response.bodyToMono(String.class)
-					.map(bodyStr -> new RuntimeException("AI 서버 오류: " + bodyStr))
-			)
-			.bodyToFlux(new ParameterizedTypeReference<ServerSentEvent<String>>() {})
-			.timeout(Duration.ofMinutes(2))
-			.doOnNext(this::logEvent);
-	}
-
 
 	//일반 메세지 통신
 	public Flux<Object> receiveSSE(final String contents, final UUID sessionId, final UUID userId) {
 		var user = validateUserAndSession(userId, sessionId);
 		AtomicReference<UUID> messageIdRef = new AtomicReference<>();
 
-		return streamSse(sessionProperties.getMessageUri(), sessionId, Map.of("content", contents))
+		return sseService.streamSSE(sessionProperties.getMessageUri(), sessionId, Map.of("content", contents))
 			.doFirst(() -> messageIdRef.set(messageService.saveContents(Role.USER, contents, sessionId)))
 			.handle((event, sink) -> handleMessageEvent(event, sink, sessionId, user, messageIdRef))
 			.doOnError(e -> log.error("스트림 오류", e));
@@ -114,7 +91,7 @@ public class SessionStreamService {
 		validateUser(userId);
 		var session = sessionRepository.findByIdOrThrow(sessionId, ErrorCode.NOT_FOUND_SESSION);
 
-		return streamSse(sessionProperties.getMessageUri(), sessionId, Map.of("content", contents))
+		return sseService.streamSSE(sessionProperties.getMessageUri(), sessionId, Map.of("content", contents))
 			.doFirst(() -> messageService.saveContents(Role.USER, contents, sessionId))
 			.handle((event, sink) ->  handleFirstMessageEvent(event, sink, session))
 			.doOnError(e -> log.error("스트림 오류", e));
@@ -163,7 +140,7 @@ public class SessionStreamService {
 		}
 		Preconditions.validate(session.getStatus().equals(SessionStatus.ACTIVE), ErrorCode.INVALID_SESSION);
 
-		return streamSse(sessionProperties.getFinalizeUri(), sessionId, null)
+		return sseService.streamSSE(sessionProperties.getFinalizeUri(), sessionId, null)
 			.handle((event, sink) -> handleFinalizeEvent(event, sink, sessionId))
 			.doOnError(e -> log.error("스트림 오류", e));
 	}
@@ -199,14 +176,6 @@ public class SessionStreamService {
 
 				messageService.saveSessionSummary(sessionId, summary, imageUrl);
 			}
-		}
-	}
-
-	private void logEvent(final ServerSentEvent<String> event) {
-		if ("error".equals(event.event())) {
-			log.error("AI error: {}", event.data());
-		} else {
-			log.info("event: {}, data: {}", event.event(), event.data()); // TODO 최종 배포 시 삭제
 		}
 	}
